Componentizing BalancerCopyIcon
