+++ packages/nextjs-app-ts/src/components/balancer-eth-components/balancer-copy-icon/IBalancerCopyIcon.ts
@@ -1,0 +1,3 @@
+export interface IBalancerCopyIcon {
+  textToCopy: string;
+}
