+++ packages/nextjs-app-ts/src/components/balancer-eth-components/balancer-generic-contract/BalancerGenericContract.tsx
@@ -1,16 +1,16 @@
-import { CopyOutlined } from '@ant-design/icons';
-import { Tooltip } from 'antd';
 import { BaseContract } from 'ethers';
 import { FC } from 'react';
 
 import { IBalancerGenericContract } from './IBalancerGenericContract';
 
+import { BalancerCopyIcon } from '~~/components/balancer-eth-components/balancer-copy-icon/BalancerCopyIcon';
 import { BalancerFunction } from '~~/components/balancer-eth-components/balancer-function/BalancerFunction';
-import { useBalancerGenericContract } from '~~/components/balancer-eth-components/balancer-generic-contract/BalancerGenericContract.hook';
 
-export const BalancerGenericContract: FC<IBalancerGenericContract<BaseContract>> = (props) => {
-  const { contract, contractAddress, contractName, copyAddressToClipboard, resetTooltip, tooltipText } =
-    useBalancerGenericContract(props);
+export const BalancerGenericContract: FC<IBalancerGenericContract<BaseContract>> = ({ contract, contractName }) => {
+  // Todo remove this (Only debugging purposes)
+  console.log(contract);
+
+  const contractAddress = contract?.address;
 
   return (
     <div className={'balancer-generic-contract-container'}>
@@ -18,13 +18,7 @@ export const BalancerGenericContract: FC<IBalancerGenericContract<BaseContract>>
         <div className={'balancer-generic-contract-name'}>{contractName}</div>
         <div className={'balancer-generic-contract-address'}>
           {contractAddress}
-          <Tooltip title={tooltipText} showArrow={false}>
-            <CopyOutlined
-              className={'balancer-copy-icon'}
-              onClick={copyAddressToClipboard}
-              onMouseLeave={resetTooltip}
-            />
-          </Tooltip>
+          <BalancerCopyIcon textToCopy={contractAddress || ''} />
         </div>
       </div>
       <div className={'balancer-generic-contract-functions'}>
